fix: Contact 타입을 전역 네임스페이스 방식으로 통일
