+++ src/fsd_entities/contact/model/hooks/index.ts
@@ -1,3 +1,2 @@
 export { useEditProfile } from '@/fsd_entities/contact/model/hooks/useEditProfile';
 export { useProfileForm } from '@/fsd_entities/contact/model/hooks/useProfileForm';
-export { useDebounce } from '@/fsd_entities/contact';
