+++ src/fsd_entities/contact/model/index.ts
@@ -1,7 +1,4 @@
 export { useUpdateMyProfileMutation } from './mutations/useUpdateMyProfileMutation';
 
 export * from './queries';
-export { useDebounce } from './hooks/useDebounce';
 export * from './hooks';
-export * from './profile.schema';
-export * from './profile.mapper';
