+++ src/fsd_shared/hooks/index.ts
@@ -4,3 +4,4 @@ export { useIsWindowLg } from './useIsWindowLg';
 export { useTruncateParagraph } from './useTruncateParagraph';
 export { useAuthHandler } from './useAuthHandler';
 export { useMoreBtnClick } from './useMoreBtnClick';
+export { useDebounce } from './useDebounce';
