+++ src/fsd_widgets/contact/ui/index.ts
@@ -1,1 +1,3 @@
 export { ContactList } from './ContactList';
+export { ContactInfoTabs } from './ContactInfoTabs';
+export { ContactDetail } from './ContactDetail';
